@@ -1,57 +1,80 @@
 {
-    "name": "fe-app",
-    "type": "module",
-    "version": "0.1.0",
-    "private": true,
-    "dependencies": {
-        "@aldabil/react-scheduler": "^2.2.1",
-        "@emotion/react": "^11.7.1",
-        "@emotion/styled": "^11.6.0",
-        "@material-ui/core": "^5.0.0-beta.5",
-        "@mui/material": "^5.3.1",
-        "@mui/x-data-grid": "^5.4.0",
-        "@mui/x-date-pickers": "^5.0.0-alpha.1",
-        "@testing-library/jest-dom": "^5.14.1",
-        "@testing-library/react": "^11.2.7",
-        "@testing-library/user-event": "^12.8.3",
-        "eslint": "^7.32.0",
-        "fetch": "^1.1.0",
-        "i18next": "^21.3.2",
-        "i18next-browser-languagedetector": "^6.1.2",
-        "i18next-xhr-backend": "^3.2.2",
-        "js-cookie": "^3.0.1",
-        "node-fetch": "^2.6.7",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-i18next": "^11.12.0",
-        "react-mui-scheduler": "^1.2.5",
-        "react-router-dom": "^5.3.0",
-        "react-scripts": "4.0.3",
-        "react-webcam": "^7.0.1",
-        "web-vitals": "^1.1.2"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject"
-    },
-    "eslintConfig": {
-        "extends": [
-            "react-app",
-            "react-app/jest"
-        ]
-    },
-    "browserslist": {
-        "production": [
-            "chrome",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    }
+  "name": "my-webpack-project",
+  "version": "1.0.0",
+  "private": true,
+  "dependencies": {
+    "@aldabil/react-scheduler": "^2.7.8",
+    "@babel/eslint-parser": "*",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/icons-material": "^5.11.11",
+    "@mui/material": "^5.11.15",
+    "@mui/x-data-grid": "^6.0.4",
+    "@mui/x-date-pickers": "^6.0.4",
+    "@rollup/plugin-terser": "^0.4.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "ajv": "^8.12.0",
+    "dayjs": "^1.11.7",
+    "fetch": "^1.1.0",
+    "i18next": "^22.4.13",
+    "i18next-browser-languagedetector": "^7.0.1",
+    "i18next-http-backend": "^2.2.0",
+    "joi": "^17.9.1",
+    "js-cookie": "^3.0.1",
+    "node": "^19.8.1",
+    "node-fetch": "^3.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.0.0",
+    "react-i18next": "^12.2.0",
+    "react-router-dom": "^6.10.0",
+    "react-webcam": "^7.0.1",
+    "typescript": "^5.0.3",
+    "web-vitals": "^3.3.0"
+  },
+  "scripts": {
+    "start": "webpack-dev-server --mode development --hot --open",
+    "build": "webpack --mode=production --node-env=production",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "build:dev": "webpack --mode=development",
+    "build:prod": "webpack --mode=production --node-env=production",
+    "watch": "webpack --watch",
+    "serve": "webpack serve"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      "last 2 chrome version",
+      "last 2 firefox version",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 2 chrome version",
+      "last 2 firefox version"
+    ]
+  },
+  "module": "",
+  "devDependencies": {
+    "@babel/core": "^7.21.4",
+    "@babel/preset-env": "^7.21.4",
+    "@babel/preset-react": "^7.18.6",
+    "@webpack-cli/generators": "^3.0.1",
+    "babel-loader": "^9.1.2",
+    "css-loader": "^6.7.3",
+    "html-webpack-plugin": "^5.5.0",
+    "prettier": "^2.8.7",
+    "style-loader": "^3.3.2",
+    "webpack": "^5.77.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.13.2"
+  },
+  "description": "My webpack project"
 }
