@@ -1,12 +1,12 @@
 import i18n from "i18next";
 import LanguageDetector from "i18next-browser-languagedetector";
-import XHR from "i18next-xhr-backend";
+import HttpApi from "i18next-http-backend";
 
 import translationEng from "../locales/en/translation.json";
 import translationEsp from "../locales/es/translation.json";
 
 i18n
-    .use(XHR)
+    .use(HttpApi)
     .use(LanguageDetector)
     .init({
         detection: { order: ["path", "navigator"] },
